move import genbank/csv into middleware as a promise returning method
