+++ src/middleware/api.js
@@ -230,6 +230,36 @@ export const infoQuery = (type, detail) => {
     .then(resp => resp.json());
 };
 
+/**
+ * import a genbank or CSV file into the given project or into a new project.
+ * project ID is returned and should be reloaded if the current project or opened if a new project.
+ * Promise resolves with projectId on success and rejects with statusText of xhr
+ */
+export const importGenbankOrCSV = (file, projectId) => {
+  invariant(file && file.name, 'expected a file object of the type that can be added to FormData');
+  const formData = new FormData();
+  formData.append('data', file, file.name);
+  const isCSV = file.name.toLowerCase().endsWith('.csv');
+  const uri = `/import/${isCSV ? 'csv' : 'genbank'}${projectId ? '/' + projectId : ''}`;
+  return new Promise((resolve, reject) => {
+    const xhr = new XMLHttpRequest();
+    xhr.open('POST', uri, true);
+    xhr.onload = () => {
+      if (xhr.status >= 200 && xhr.status < 300) {
+        const json = JSON.parse(xhr.response);
+        invariant(json && json.ProjectId, 'expect a project ID');
+        resolve(json.ProjectId);
+      } else {
+        reject(xhr.statusText);
+      }
+    }
+    xhr.onerror = () => {
+      reject(xhr.statusText);
+    }
+    xhr.send(formData);
+  });
+};
+
 /*************************
  Sequence API
  *************************/
